+++ lib/mock_command.dart
@@ -3,13 +3,64 @@ part of './command_it.dart';
 /// `MockCommand` allows you to easily mock an Command for your Unit and UI tests
 /// Mocking a command with `mockito` https://pub.dartlang.org/packages/mockito has its limitations.
 class MockCommand<TParam, TResult> extends Command<TParam, TResult> {
-  List<CommandResult<TParam, TResult>>? returnValuesForNextExecute;
+  /// Internal storage for queued results - use [queueResultsForNextRunCall] to set
+  List<CommandResult<TParam, TResult>>? returnValuesForNextRun;
 
   /// the last value that was passed when run or the command directly was called
-  TParam? lastPassedValueToExecute;
+  TParam? lastPassedValueToRun;
 
   /// Number of times run or the command directly was called
-  int executionCount = 0;
+  int runCount = 0;
+
+  /// Deprecated: Use [runCount] instead.
+  /// This property will be removed in v10.0.0.
+  @Deprecated(
+    'Use runCount instead. '
+    'This will be removed in v10.0.0. '
+    'See BREAKING_CHANGE_EXECUTE_TO_RUN.md for migration guide.',
+  )
+  int get executionCount => runCount;
+
+  @Deprecated(
+    'Use runCount instead. '
+    'This will be removed in v10.0.0. '
+    'See BREAKING_CHANGE_EXECUTE_TO_RUN.md for migration guide.',
+  )
+  set executionCount(int value) => runCount = value;
+
+  /// Deprecated: Use [lastPassedValueToRun] instead.
+  /// This property will be removed in v10.0.0.
+  @Deprecated(
+    'Use lastPassedValueToRun instead. '
+    'This will be removed in v10.0.0. '
+    'See BREAKING_CHANGE_EXECUTE_TO_RUN.md for migration guide.',
+  )
+  TParam? get lastPassedValueToExecute => lastPassedValueToRun;
+
+  @Deprecated(
+    'Use lastPassedValueToRun instead. '
+    'This will be removed in v10.0.0. '
+    'See BREAKING_CHANGE_EXECUTE_TO_RUN.md for migration guide.',
+  )
+  set lastPassedValueToExecute(TParam? value) => lastPassedValueToRun = value;
+
+  /// Deprecated: Use [returnValuesForNextRun] instead.
+  /// This property will be removed in v10.0.0.
+  @Deprecated(
+    'Use returnValuesForNextRun instead. '
+    'This will be removed in v10.0.0. '
+    'See BREAKING_CHANGE_EXECUTE_TO_RUN.md for migration guide.',
+  )
+  List<CommandResult<TParam, TResult>>? get returnValuesForNextExecute =>
+      returnValuesForNextRun;
+
+  @Deprecated(
+    'Use returnValuesForNextRun instead. '
+    'This will be removed in v10.0.0. '
+    'See BREAKING_CHANGE_EXECUTE_TO_RUN.md for migration guide.',
+  )
+  set returnValuesForNextExecute(List<CommandResult<TParam, TResult>>? value) =>
+      returnValuesForNextRun = value;
 
   /// constructor that can take an optional `ValueListenable` to control if the command can be run
   /// if the wrapped function has `void` as return type [noResult] has to be `true`
@@ -36,17 +87,23 @@ class MockCommand<TParam, TResult> extends Command<TParam, TResult> {
         .listen((result, _) => value = result.data!);
   }
 
-  /// to be able to simulate any output of the command when it is called you can here queue the output data for the next execution call
+  /// Deprecated: Use [queueResultsForNextRunCall] instead.
+  /// This method will be removed in v10.0.0.
+  @Deprecated(
+    'Use queueResultsForNextRunCall() instead. '
+    'This will be removed in v10.0.0. '
+    'See BREAKING_CHANGE_EXECUTE_TO_RUN.md for migration guide.',
+  )
   // ignore: use_setters_to_change_properties
   void queueResultsForNextExecuteCall(
     List<CommandResult<TParam, TResult>> values,
   ) {
-    returnValuesForNextExecute = values;
+    queueResultsForNextRunCall(values);
   }
 
   /// Can either be called directly or by calling the object itself because Commands are callable classes
-  /// Will increase [executionCount] and assign [lastPassedValueToExecute] the value of [param]
-  /// If you have queued a result with [queueResultsForNextExecuteCall] it will be copies tho the output stream.
+  /// Will increase [runCount] and assign [lastPassedValueToRun] the value of [param]
+  /// If you have queued a result with [queueResultsForNextRunCall] it will be copies tho the output stream.
   /// [isRunning], [canRun] and [results] will work as with a real command.
   @override
   void run([TParam? param]) {
@@ -59,12 +116,12 @@ class MockCommand<TParam, TResult> extends Command<TParam, TResult> {
     }
 
     _isRunning.value = true;
-    executionCount++;
-    lastPassedValueToExecute = param;
+    runCount++;
+    lastPassedValueToRun = param;
     // ignore: avoid_print
     print('Called Execute');
-    if (returnValuesForNextExecute != null) {
-      returnValuesForNextExecute!.map((entry) {
+    if (returnValuesForNextRun != null) {
+      returnValuesForNextRun!.map((entry) {
         if ((entry.isRunning || entry.hasError) &&
             _includeLastResultInCommandResults) {
           return CommandResult<TParam, TResult>(
@@ -95,14 +152,58 @@ class MockCommand<TParam, TResult> extends Command<TParam, TResult> {
   @override
   void execute([TParam? param]) => run(param);
 
+  /// Deprecated: Use [startRun] instead.
+  /// This method will be removed in v10.0.0.
+  @Deprecated(
+    'Use startRun() instead. '
+    'This will be removed in v10.0.0. '
+    'See BREAKING_CHANGE_EXECUTE_TO_RUN.md for migration guide.',
+  )
+  void startExecution([TParam? param]) {
+    startRun(param);
+  }
+
+  /// Deprecated: Use [endRunWithData] instead.
+  /// This method will be removed in v10.0.0.
+  @Deprecated(
+    'Use endRunWithData() instead. '
+    'This will be removed in v10.0.0. '
+    'See BREAKING_CHANGE_EXECUTE_TO_RUN.md for migration guide.',
+  )
+  void endExecutionWithData(TResult data) {
+    endRunWithData(data);
+  }
+
+  /// Deprecated: Use [endRunWithError] instead.
+  /// This method will be removed in v10.0.0.
+  @Deprecated(
+    'Use endRunWithError() instead. '
+    'This will be removed in v10.0.0. '
+    'See BREAKING_CHANGE_EXECUTE_TO_RUN.md for migration guide.',
+  )
+  void endExecutionWithError(String message) {
+    endRunWithError(message);
+  }
+
+  /// Deprecated: Use [endRunNoData] instead.
+  /// This method will be removed in v10.0.0.
+  @Deprecated(
+    'Use endRunNoData() instead. '
+    'This will be removed in v10.0.0. '
+    'See BREAKING_CHANGE_EXECUTE_TO_RUN.md for migration guide.',
+  )
+  void endExecutionNoData() {
+    endRunNoData();
+  }
+
   /// For a more fine grained control to simulate the different states of an [Command]
   /// there are these functions
-  /// `startExecution` will issue a [CommandResult] with
+  /// `startRun` will issue a [CommandResult] with
   /// data: null
   /// error: null
-  /// isExecuting : true
-  void startExecution([TParam? param]) {
-    lastPassedValueToExecute = param;
+  /// isRunning : true
+  void startRun([TParam? param]) {
+    lastPassedValueToRun = param;
     _commandResult.value = CommandResult<TParam, TResult>(
       param,
       _includeLastResultInCommandResults ? value : null,
@@ -112,14 +213,14 @@ class MockCommand<TParam, TResult> extends Command<TParam, TResult> {
     _isRunning.value = true;
   }
 
-  /// `endExecutionWithData` will issue a [CommandResult] with
+  /// `endRunWithData` will issue a [CommandResult] with
   /// data: [data]
   /// error: null
-  /// isExecuting : false
-  void endExecutionWithData(TResult data) {
+  /// isRunning : false
+  void endRunWithData(TResult data) {
     value = data;
     _commandResult.value = CommandResult<TParam, TResult>(
-      lastPassedValueToExecute,
+      lastPassedValueToRun,
       data,
       null,
       false,
@@ -130,13 +231,13 @@ class MockCommand<TParam, TResult> extends Command<TParam, TResult> {
     _isRunning.value = false;
   }
 
-  /// `endExecutionWithData` will issue a [CommandResult] with
+  /// `endRunWithError` will issue a [CommandResult] with
   /// data: null
   /// error: Exception([message])
-  /// isExecuting : false
-  void endExecutionWithError(String message) {
+  /// isRunning : false
+  void endRunWithError(String message) {
     _handleErrorFiltered(
-      lastPassedValueToExecute,
+      lastPassedValueToRun,
       Exception(message),
       StackTrace.current,
     );
@@ -146,13 +247,13 @@ class MockCommand<TParam, TResult> extends Command<TParam, TResult> {
     }
   }
 
-  /// `endExecutionNoData` will issue a [CommandResult] with
+  /// `endRunNoData` will issue a [CommandResult] with
   /// data: null
   /// error: null
-  /// isExecuting : false
-  void endExecutionNoData() {
+  /// isRunning : false
+  void endRunNoData() {
     _commandResult.value = CommandResult<TParam, TResult>(
-      lastPassedValueToExecute,
+      lastPassedValueToRun,
       _includeLastResultInCommandResults ? value : null,
       null,
       false,
@@ -161,6 +262,14 @@ class MockCommand<TParam, TResult> extends Command<TParam, TResult> {
       Command.loggingHandler?.call(_name, _commandResult.value);
     }
     _isRunning.value = false;
+  }
+
+  /// to be able to simulate any output of the command when it is called you can here queue the output data for the next run call
+  // ignore: use_setters_to_change_properties
+  void queueResultsForNextRunCall(
+    List<CommandResult<TParam, TResult>> values,
+  ) {
+    returnValuesForNextRun = values;
   }
 
   @override